automatically generate diagram if needed when opening a *.pi refinement
